+++ lib/src/state_machine/mp_th2_file_edit_state_machine/mp_th2_file_edit_state_select_empty_selection.dart
@@ -73,7 +73,7 @@ class MPTH2FileEditStateSelectEmptySelection extends MPTH2FileEditState
         if ((isCtrlPressed || isMetaPressed) &&
             !isAltPressed &&
             !isShiftPressed) {
-          th2FileEditController.elementEditController.pasteElements();
+          th2FileEditController.copyPasteController.pasteElements();
           keyProcessed = true;
         }
     }