+++ test/3158-ui_duplicate_area_ctrl_d_test.dart
@@ -100,7 +100,7 @@ void main() {
       th2Controller.stateController.setState(
         MPTH2FileEditStateType.selectNonEmptySelection,
       );
-      th2Controller.elementEditController.duplicateSelectedElements();
+      th2Controller.copyPasteController.duplicateSelectedElements();
 
       final List<THArea> areasAfterDuplicate = th2File.getAreas().toList();
       final List<THLine> linesAfterDuplicate = th2File.getLines().toList();
@@ -204,7 +204,7 @@ void main() {
       th2Controller.stateController.setState(
         MPTH2FileEditStateType.selectNonEmptySelection,
       );
-      th2Controller.elementEditController.duplicateSelectedElements();
+      th2Controller.copyPasteController.duplicateSelectedElements();
 
       final List<THArea> areasAfterDuplicate = th2File.getAreas().toList();
       final List<THLine> linesAfterDuplicate = th2File.getLines().toList();