+++ test/3164-ui_copy_paste_multi_elements_same_scrap_test.dart
@@ -122,7 +122,7 @@ void main() {
           );
 
           /// Copy all selected elements
-          th2Controller.elementEditController.copySelectedElements();
+          th2Controller.copyPasteController.copySelectedElements();
 
           /// Verify clipboard has content
           expect(mpLocator.mpGeneralController.hasClipboardContent, isTrue);
@@ -131,7 +131,7 @@ void main() {
           th2Controller.stateController.setState(
             MPTH2FileEditStateType.selectEmptySelection,
           );
-          th2Controller.elementEditController.pasteElements();
+          th2Controller.copyPasteController.pasteElements();
 
           /// Verify all elements were pasted
           final List<THPoint> pointsAfterPaste = th2File.getPoints().toList();